fix deadlock with R generators
with keras 2.13, generators are
converted to tf.data pipelines with
a prefetch operation. This caused a deadlock,
where a queue consumer for the generator was
not insync with the main-thread queue producer
scheduled as a keras callback. This removes the callback scheduler
and manual eventloop, defaulting to standard python evaluation semantics,
where we depend on tasks scheduled on the main thread to be able to run
while the main thread is blocked.

related: https://github.com/rstudio/tensorflow/issues/575